+++ Magic/CHANGELOG.md
@@ -38,6 +38,7 @@
  - Fix a whole pile of memory leaks
  - Improved support for persistent mobs (in 1.14 and up)
  - Add "land" trigger and "fall_distance" attribute
+ - Allow ChangeContext sourceLocation and targetLocation parameters to specify feet, head, etc entity locations
 
 # 8.1.2